refactor DeepTutorChat
