+++ chrome/content/zotero/DeepTutorTopSection.js
@@ -1,6 +1,6 @@
 import React from 'react'; // eslint-disable-line no-unused-vars
 
-// Top section removed: buttons moved into chat header within DeepTutorChatBox.
+// Top section removed: buttons moved into chat header within DeepTutorChat.
 // Keeping a no-op component to avoid import breakages.
 const DeepTutorTopSection = () => null;
